requirements.txt
Pillow==9.5.0
pandas==1.4.2
GitPython==3.1.32
matplotlib==3.7.0
numpy==1.23.5
opencv-python==4.7.0.72
psutil==5.9.4
PyYAML==6.0
requests==2.28.2
scipy==1.9.1
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
ultralytics==8.0.149
scikit-learn==1.2.2
loguru==0.6.0
tensorboard==2.11.2
wandb==0.15.8
pycocotools==2.0.6
albumentations==1.3.1
vision-transformers==0.1.1.0
torchinfo==1.8.0
catboost
umap-learn
yacs
wget